Fix migrations not running on live db during deploy.
I don’t stagehand is fully configured when we check if we are in single_connection mode in this task. Once Rails is fully loaded, we aren’t in single connection mode but it is too late and these rake tasks are no longer enhanced.

We already do `Stagehand::Database.each` in the task so this check seems unnecessary.